+++ backend/ensembles/serializers/arrangement.py
@@ -1,0 +1,39 @@
+from rest_framework import serializers
+
+from ensembles.models import Arrangement
+
+from ensembles.serializers.arrangement_version import ArrangementVersionSerializer
+
+from logging import getLogger
+
+logger = getLogger("EnsembleViews")
+
+
+VERSION_TYPES = [("major", "Major"), ("minor", "Minor"), ("patch", "Patch")]
+
+
+class ArrangementSerializer(serializers.ModelSerializer):
+    latest_version = ArrangementVersionSerializer(read_only=True)
+    latest_version_num = serializers.ReadOnlyField()
+
+    class Meta:
+        model = Arrangement
+        fields = [
+            "id",
+            "ensemble",
+            "ensemble_name",
+            "ensemble_slug",
+            "title",
+            "slug",
+            "subtitle",
+            "composer",
+            "mvt_no",
+            "style",
+            "latest_version",
+            "latest_version_num",
+        ]
+        read_only_fields = [
+            "slug",
+            "ensemble_name",
+            "ensemble_slug",
+        ]
