+++ backend/ensembles/serializers/arrangement_version.py
@@ -1,0 +1,104 @@
+from rest_framework import serializers
+from rest_framework import status
+
+from django.db import transaction
+from django.core.files.storage import default_storage
+
+from ensembles.models import Arrangement, ArrangementVersion
+from ensembles.tasks import prep_and_export_mscz, export_arrangement_version
+
+from logging import getLogger
+
+logger = getLogger("EnsembleViews")
+
+
+VERSION_TYPES = [("major", "Major"), ("minor", "Minor"), ("patch", "Patch")]
+
+
+class ArrangementVersionSerializer(serializers.ModelSerializer):
+    audio_state = serializers.CharField(
+        source="get_audio_state_display", read_only=True
+    )
+
+    class Meta:
+        model = ArrangementVersion
+        fields = ["id", "version_label", "timestamp", "is_latest", "audio_state"]
+
+
+
+class CreateArrangementVersionMsczSerializer(serializers.Serializer):
+    default_error_messages = {
+        "invalid_version_type": "Version type not one of 'major', 'minor', 'patch'",
+        "invalid_arr_id": "Arrangement with id {id} does not exist",
+    }
+
+    file = serializers.FileField(allow_empty_file=False)
+    arrangement_id = serializers.IntegerField(required=True)
+    version_type = serializers.CharField(required=True)
+    num_measures_per_line_score = serializers.IntegerField(default=8)
+    num_measures_per_line_part = serializers.IntegerField(default=6)
+    num_lines_per_page = serializers.IntegerField(default=8)
+
+    format_parts = serializers.BooleanField(default=True)
+
+    def validate_version_type(self, value):
+        if value not in [t[0] for t in VERSION_TYPES]:
+            self.fail("invalid_version_type")
+        return value
+
+    def validate_arrangement_id(self, value):
+        if not Arrangement.objects.filter(id=value).exists():
+            self.fail("invalid_arr_id", id=value)
+        return value
+
+    def save(self, **kwargs):
+        assert self.validated_data, "Must call is_valid first!"
+        with transaction.atomic():
+            version = ArrangementVersion.objects.create(
+                arrangement=Arrangement.objects.get(
+                    id=self.validated_data["arrangement_id"]
+                ),
+                file_name=self.validated_data["file"].name,
+                num_measures_per_line_score=self.validated_data[
+                    "num_measures_per_line_score"
+                ],
+                num_measures_per_line_part=self.validated_data[
+                    "num_measures_per_line_part"
+                ],
+                num_lines_per_page=self.validated_data["num_lines_per_page"],
+            )
+
+            version.save(
+                version_type=self.validated_data["version_type"],
+            )
+
+        uploaded_file = self.validated_data["file"]
+
+        # Save file to storage using the storage key
+        try:
+            # Create a file-like object from the uploaded file
+            file_content = b""
+            for chunk in uploaded_file.chunks():
+                file_content += chunk
+
+            # Save to storage using the key
+            default_storage.save(version.mscz_file_key, io.BytesIO(file_content))
+            logger.info(f"Saved file to storage: {version.mscz_file_key}")
+
+        except Exception as e:
+            logger.error(f"Failed to save file to storage: {e}")
+            # Clean up the version if file save failed
+            version.delete()
+            return {"error": "Failed to save file to storage"}
+
+        # Format mscz if selected by FE
+        if self.validated_data.get("format_parts", None):
+            prep_and_export_mscz.delay(version.pk)
+        else:
+            # Just export
+            export_arrangement_version.delay(version.pk)
+
+        # epxort MXL for diff calculation (unnecessary)
+        # export_arrangement_version(version.pk, action="mxl")
+
+        return {"success": True, "version_id": version.id}
