+++ backend/ensembles/serializers/ensemble.py
@@ -1,15 +1,11 @@
-import io
 from rest_framework import serializers
-from rest_framework import status
 
-from django.db import transaction
 from django.core.files.storage import default_storage
 from django.conf import settings
 
-from ensembles.models import Ensemble, EnsembleUsership, Arrangement, ArrangementVersion, PartBook, PartName
-from ensembles.tasks import prep_and_export_mscz, export_arrangement_version
-
-from django.contrib.auth import get_user_model
+from ensembles.models import Ensemble, EnsembleUsership, PartBook, PartName
+from ensembles.serializers.arrangement import ArrangementSerializer
+from ensembles.serializers.ensemble_usership import EnsembleUsershipSerializer
 
 from logging import getLogger
 
@@ -17,43 +13,6 @@ logger = getLogger("EnsembleViews")
 
 
 VERSION_TYPES = [("major", "Major"), ("minor", "Minor"), ("patch", "Patch")]
-
-
-class ArrangementVersionSerializer(serializers.ModelSerializer):
-    audio_state = serializers.CharField(
-        source="get_audio_state_display", read_only=True
-    )
-
-    class Meta:
-        model = ArrangementVersion
-        fields = ["id", "version_label", "timestamp", "is_latest", "audio_state"]
-
-
-class ArrangementSerializer(serializers.ModelSerializer):
-    latest_version = ArrangementVersionSerializer(read_only=True)
-    latest_version_num = serializers.ReadOnlyField()
-
-    class Meta:
-        model = Arrangement
-        fields = [
-            "id",
-            "ensemble",
-            "ensemble_name",
-            "ensemble_slug",
-            "title",
-            "slug",
-            "subtitle",
-            "composer",
-            "mvt_no",
-            "style",
-            "latest_version",
-            "latest_version_num",
-        ]
-        read_only_fields = [
-            "slug",
-            "ensemble_name",
-            "ensemble_slug",
-        ]
 
 
 class EnsembleSerializer(serializers.ModelSerializer):
@@ -116,13 +75,7 @@ class EnsembleSerializer(serializers.ModelSerializer):
         request = self.context.get("request")
         if request and request.user.is_authenticated:
             return [
-                #TODO[SC-278]: Usership serializer
-                {
-                    "id": usership.id,
-                    "user": UserSerializer(usership.user).data,
-                    "role": usership.role,
-                    "date_joined": usership.date_joined,
-                }
+                EnsembleUsershipSerializer(usership).data
                 for usership in EnsembleUsership.objects.filter(ensemble=obj)
             ]
         return None
@@ -150,6 +103,7 @@ class EnsembleSerializer(serializers.ModelSerializer):
         )
         result = []
         for book in books:
+            #TODO[SC-278]: Serializer
             item = {
                 "id": book.id,
                 "part_name_id": book.part_name_id,
@@ -205,92 +159,3 @@ class EnsemblePartNameMergeSerializer(serializers.Serializer):
         attrs["second_part"] = second_part
 
         return attrs
-
-        
-
-
-
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = get_user_model()
-        fields = ["id", "username", "email"]
-
-
-class CreateArrangementVersionMsczSerializer(serializers.Serializer):
-    default_error_messages = {
-        "invalid_version_type": "Version type not one of 'major', 'minor', 'patch'",
-        "invalid_arr_id": "Arrangement with id {id} does not exist",
-    }
-
-    file = serializers.FileField(allow_empty_file=False)
-    arrangement_id = serializers.IntegerField(required=True)
-    version_type = serializers.CharField(required=True)
-    num_measures_per_line_score = serializers.IntegerField(default=8)
-    num_measures_per_line_part = serializers.IntegerField(default=6)
-    num_lines_per_page = serializers.IntegerField(default=8)
-
-    format_parts = serializers.BooleanField(default=True)
-
-    def validate_version_type(self, value):
-        if value not in [t[0] for t in VERSION_TYPES]:
-            self.fail("invalid_version_type")
-        return value
-
-    def validate_arrangement_id(self, value):
-        if not Arrangement.objects.filter(id=value).exists():
-            self.fail("invalid_arr_id", id=value)
-        return value
-
-    def save(self, **kwargs):
-        assert self.validated_data, "Must call is_valid first!"
-        with transaction.atomic():
-            version = ArrangementVersion.objects.create(
-                arrangement=Arrangement.objects.get(
-                    id=self.validated_data["arrangement_id"]
-                ),
-                file_name=self.validated_data["file"].name,
-                num_measures_per_line_score=self.validated_data[
-                    "num_measures_per_line_score"
-                ],
-                num_measures_per_line_part=self.validated_data[
-                    "num_measures_per_line_part"
-                ],
-                num_lines_per_page=self.validated_data["num_lines_per_page"],
-            )
-
-            version.save(
-                version_type=self.validated_data["version_type"],
-            )
-
-        uploaded_file = self.validated_data["file"]
-
-        # Save file to storage using the storage key
-        try:
-            # Create a file-like object from the uploaded file
-            file_content = b""
-            for chunk in uploaded_file.chunks():
-                file_content += chunk
-
-            # Save to storage using the key
-            default_storage.save(version.mscz_file_key, io.BytesIO(file_content))
-            logger.info(f"Saved file to storage: {version.mscz_file_key}")
-
-        except Exception as e:
-            logger.error(f"Failed to save file to storage: {e}")
-            # Clean up the version if file save failed
-            version.delete()
-            return {"error": "Failed to save file to storage"}
-
-        # Format mscz if selected by FE
-        if self.validated_data.get("format_parts", None):
-            prep_and_export_mscz.delay(version.pk)
-        else:
-            # Just export
-            export_arrangement_version.delay(version.pk)
-
-        # epxort MXL for diff calculation (unnecessary)
-        # export_arrangement_version(version.pk, action="mxl")
-
-        return {"success": True, "version_id": version.id}
